# storage/diskmap: remove SortedDiskMapIterator.{Key,Value}