--- a/pkg/sql/rowcontainer/disk_row_container.go
+++ b/pkg/sql/rowcontainer/disk_row_container.go
@@ -275,6 +275,7 @@ func (d *DiskRowContainer) keyValToRow(k []byte, v []byte) (sqlbase.EncDatumRow,
 // diskRowIterator iterates over the rows in a DiskRowContainer.
 type diskRowIterator struct {
 	rowContainer *DiskRowContainer
+	rowBuf       []byte
 	diskmap.SortedDiskMapIterator
 }
 
@@ -305,7 +306,23 @@ func (r *diskRowIterator) Row() (sqlbase.EncDatumRow, error) {
 		return nil, errors.AssertionFailedf("invalid row")
 	}
 
-	return r.rowContainer.keyValToRow(r.Key(), r.Value())
+	k := r.UnsafeKey()
+	v := r.UnsafeValue()
+	// TODO(asubiotto): the "true ||" should not be necessary. We should be to
+	// reuse rowBuf, yet doing so causes
+	// TestDiskBackedIndexedRowContainer/ReorderingOnDisk, TestHashJoiner, and
+	// TestSorter to fail. Some caller of Row() is presumably not making a copy
+	// of the return value.
+	if true || cap(r.rowBuf) < len(k)+len(v) {
+		r.rowBuf = make([]byte, 0, len(k)+len(v))
+	}
+	r.rowBuf = r.rowBuf[:len(k)+len(v)]
+	copy(r.rowBuf, k)
+	copy(r.rowBuf[len(k):], v)
+	k = r.rowBuf[:len(k)]
+	v = r.rowBuf[len(k):]
+
+	return r.rowContainer.keyValToRow(k, v)
 }
 
 func (r *diskRowIterator) Close() {
@@ -347,7 +364,8 @@ func (r *diskRowFinalIterator) Row() (sqlbase.EncDatumRow, error) {
 	if err != nil {
 		return nil, err
 	}
-	r.diskRowIterator.rowContainer.lastReadKey = r.Key()
+	r.diskRowIterator.rowContainer.lastReadKey =
+		append(r.diskRowIterator.rowContainer.lastReadKey[:0], r.UnsafeKey()...)
 	return row, nil
 }